+++ NEWS.md
@@ -10,6 +10,9 @@
 * `gghighlight()` now ignores if the calculation fails over some layers. This
   is useful to combine with such layers as `annotate()` (#78).
 
+* `gghighlight()` now allows to highlight 0-layer plots, which means just
+  filtering the plot data (#81).
+
 * `gghighlight()` now ignores `NA`s in numeric predicates (#86).
 
 # gghighlight 0.1.0